+++ helpers/report.sh
@@ -1,0 +1,4 @@
+drydock uat --report Toml
+drydock uat --report jq
+drydock uat --report CommonMark
+drydock uat --report ReadingList